requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
packaging==23.1
promise==2.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
